requirements.txt
Django>=5.1,<6.0
django-allauth>=65.0.0
requests>=2.31.0
